+++ components/bridge/api/src/main/java/com/flipperdevices/bridge/api/manager/service/FlipperVersionApi.kt
@@ -1,8 +1,8 @@
 package com.flipperdevices.bridge.api.manager.service
 
-import com.flipperdevices.bridge.api.model.FlipperVersionInformation
+import com.flipperdevices.bridge.api.model.SemVer
 import kotlinx.coroutines.flow.StateFlow
 
 interface FlipperVersionApi {
-    fun getVersionInformationFlow(): StateFlow<FlipperVersionInformation?>
+    fun getVersionInformationFlow(): StateFlow<SemVer?>
 }